+++ versions.tf
@@ -26,7 +26,7 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-bigquery/v6.0.0"
+    module_name = "blueprints/terraform/terraform-google-bigquery/v6.1.1"
   }
 
 }